requirements.txt
twilio
openai
elevenlabs
pydub
flask
soundfile
python-dotenv
